feat: restructure application layout and add styling assets
