+++ quick-loan-ui/src/assets/styles/App.css
