+++ quick-loan-ui/src/assets/styles/index.css
